apy-tracking move into functions
