--- packages/bots/apy-tracking/index.ts
@@ -1,8 +1,0 @@
-import { updateFlyWheelData, updatePluginsData } from './controllers';
-
-const main = async () => {
-  await updatePluginsData();
-  await updateFlyWheelData();
-};
-
-main();
