--- packages/bots/apy-tracking/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@midas-capital/apy-tracking",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "ts-node index.ts"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@midas-capital/sdk": "workspace:^",
-    "@supabase/supabase-js": "^1.35.4",
-    "dotenv": "^16.0.1",
-    "ethers": "^5.6.9",
-    "ts-node": "^10.8.2"
-  }
-}
