+++ packages/functions/src/config/variables.ts
